Fix issue with query flags
Query flags are the Http equivalent of booleans. You can either set them
or you cannot. Currently the generated ruby code treats them as values:
You pass in a string and it gets added to the query string.

This ensures query flags in a servant API are modeled as booleans in the
generated ruby code.