Query flags result in optional args
When adding a query flag to an existing endpoint it would be nice if
this were a backwards-compatible change. From the HTTP perspective it
already is: the query flag is an optional element of the URI that can be
omitted. But code generation produces a non-optional argument. This
imposes a hard order-of-deploy between the service that supports the new
parameter and the client that makes use of it (or not).

This changes the code so the query flag is an optional argument in ruby
generated code. That way the backwards-compatibility of adding a query
parameter to an API is preserved into our generated code.